#11 - Move `SearchDrawer` components into main folder
