+++ src/pages/Main/SearchDrawer/SkeletonResults.tsx
